+++ packages/swingset/src/lib/registry.ts
@@ -50,6 +50,7 @@ import {
   Default as OrganizationProfileGeneralDefault,
   meta as organizationProfileGeneralMeta,
 } from '../stories/organization-profile-general.stories';
+import { meta as otpMeta } from '../stories/otp.stories';
 import { meta as popoverMeta } from '../stories/popover.stories';
 import { meta as selectMeta } from '../stories/select.stories';
 import { Default as TabsComponentDefault, meta as tabsComponentMeta } from '../stories/tabs.component.stories';
@@ -111,6 +112,7 @@ const dialogModule: StoryModule = { meta: dialogMeta };
 const drawerModule: StoryModule = { meta: drawerMeta };
 const fileUploadModule: StoryModule = { meta: fileUploadMeta };
 const menuModule: StoryModule = { meta: menuMeta };
+const otpModule: StoryModule = { meta: otpMeta };
 const popoverModule: StoryModule = { meta: popoverMeta };
 const selectModule: StoryModule = { meta: selectMeta };
 const tabsModule: StoryModule = { meta: tabsMeta };
@@ -145,6 +147,7 @@ export const registry: StoryModule[] = [
   drawerModule,
   fileUploadModule,
   menuModule,
+  otpModule,
   popoverModule,
   selectModule,
   tabsModule,